+++ examples/full/main.go
@@ -18,7 +18,7 @@ import (
 	"github.com/lcylpzls/authx/session"
 	"github.com/lcylpzls/authx/token"
 	"github.com/lcylpzls/logx"
-	"github.com/lcylpzls/webx/v2"
+	"github.com/lcylpzls/webx"
 )
 
 func main() {